[#12] Feat: add all entity
