+++ src/main/java/com/example/dp/domain/like/entity/Like.java
@@ -1,5 +1,8 @@
-package com.example.dp.domain.menu.entity;
+package com.example.dp.domain.like.entity;
 
+import com.example.dp.domain.menu.entity.Menu;
+import com.example.dp.domain.model.TimeEntity;
+import com.example.dp.domain.user.entity.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -9,30 +12,24 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
+@Table(name = "TB_LIKE")
 @Getter
-@Table(name = "TB_MENU_CATEGORY")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MenuCategory {
+public class Like extends TimeEntity {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "category_id", nullable = false)
-    private Category category;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "menu_id", nullable = false)
+    @JoinColumn(name = "menu_id")
     private Menu menu;
 
-    @Builder
-    private MenuCategory(Category category, Menu menu){
-        this.category =  category;
-        this.menu = menu;
-    }
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
 }
