+++ src/main/java/com/example/dp/domain/menucategory/entity/MenuCategory.java
@@ -1,10 +1,7 @@
-package com.example.dp.domain.cart.entity;
+package com.example.dp.domain.menucategory.entity;
 
-
+import com.example.dp.domain.category.entity.Category;
 import com.example.dp.domain.menu.entity.Menu;
-import com.example.dp.domain.model.TimeEntity;
-import com.example.dp.domain.user.entity.User;
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -12,38 +9,33 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Getter
 @Entity
-@Table(name = "TB_CART")
+@Getter
+@Table(name = "TB_MENU_CATEGORY")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Cart extends TimeEntity {
+public class MenuCategory {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @OneToOne
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "category_id", nullable = false)
+    private Category category;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "menu_id", nullable = false)
     private Menu menu;
 
-    @Column(nullable = false)
-    private Long menuCount;
-
     @Builder
-    private Cart(User user, Menu menu, Long menuCount) {
-        this.user = user;
+    private MenuCategory(Category category, Menu menu) {
+        this.category = category;
         this.menu = menu;
-        this.menuCount = menuCount;
     }
 }
