+++ src/main/java/com/example/dp/domain/order/entity/Order.java
@@ -1,49 +1,37 @@
-package com.example.dp.domain.cart.entity;
+package com.example.dp.domain.order.entity;
 
-
-import com.example.dp.domain.menu.entity.Menu;
 import com.example.dp.domain.model.TimeEntity;
 import com.example.dp.domain.user.entity.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Getter
 @Entity
-@Table(name = "TB_CART")
+@Table(name = "TB_ORDER")
+@Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Cart extends TimeEntity {
+public class Order extends TimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @OneToOne
-    @JoinColumn(name = "user_id", nullable = false)
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
     private User user;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "menu_id", nullable = false)
-    private Menu menu;
-
     @Column(nullable = false)
-    private Long menuCount;
-
-    @Builder
-    private Cart(User user, Menu menu, Long menuCount) {
-        this.user = user;
-        this.menu = menu;
-        this.menuCount = menuCount;
-    }
+    @Enumerated(EnumType.STRING)
+    private OrderStatus status;
 }
