+++ src/main/java/com/example/dp/domain/order/entity/OrderStatus.java
@@ -1,0 +1,5 @@
+package com.example.dp.domain.order.entity;
+
+public enum OrderStatus {
+
+}
