+++ src/main/java/com/example/dp/domain/user/UserRole.java
@@ -13,6 +13,7 @@ public enum UserRole {
     private final String authority;
 
     public static class Authority {
+
         public static final String USER = "ROLE_USER";
         public static final String ADMIN = "ROLE_ADMIN";
     }
